# api/openapi.yaml
openapi: 3.0.3
info:
  title: key-agent API
  description: Local Key-Value and Secrets Management API
  version: 1.0.0
  contact:
    name: key-agent
    url: https://github.com/key-agent/key-agent

servers:
  - url: http://127.0.0.1:8080
    description: Local server

security:
  - bearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  version:
                    type: string
                    example: 1.0.0

  /api/v1/kv:
    get:
      summary: List all KV keys
      operationId: listKV
      parameters:
        - name: prefix
          in: query
          description: Filter keys by prefix
          schema:
            type: string
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/kv/{key}:
    get:
      summary: Get a KV entry
      operationId: getKV
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: KV entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Set a KV entry
      operationId: setKV
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetKVRequest'
      responses:
        '200':
          description: KV entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '201':
          description: KV entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: Delete a KV entry
      operationId: deleteKV
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Entry deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/secrets:
    get:
      summary: List all secret keys
      operationId: listSecrets
      parameters:
        - name: prefix
          in: query
          description: Filter keys by prefix
          schema:
            type: string
      responses:
        '200':
          description: List of secret keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/secrets/{key}:
    get:
      summary: Get a secret entry
      operationId: getSecret
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Secret entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Set a secret entry
      operationId: setSecret
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSecretRequest'
      responses:
        '200':
          description: Secret entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretEntry'
        '201':
          description: Secret entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: Delete a secret entry
      operationId: deleteSecret
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Entry deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/token:
    post:
      summary: Create a new access token
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        '201':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Access token

  schemas:
    Metadata:
      type: object
      properties:
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        expires_at:
          type: string
          format: date-time

    Entry:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
        value:
          type: string
          description: The stored value (decrypted)
        metadata:
          $ref: '#/components/schemas/Metadata'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64

    SecretEntry:
      allOf:
        - $ref: '#/components/schemas/Entry'
        - type: object
          required:
            - type
          properties:
            type:
              type: string
              enum:
                - password
                - api_key
                - certificate
                - private_key
                - token
                - other

    SetKVRequest:
      type: object
      required:
        - value
      properties:
        value:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'

    SetSecretRequest:
      type: object
      required:
        - value
        - type
      properties:
        value:
          type: string
        type:
          type: string
          enum:
            - password
            - api_key
            - certificate
            - private_key
            - token
            - other
        metadata:
          $ref: '#/components/schemas/Metadata'

    CreateTokenRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Token name for identification
        type:
          type: string
          enum:
            - client
            - mcp
          default: client
          description: Token type
        expires_in:
          type: string
          description: Duration until expiration (e.g., "24h", "30d")

    CreateTokenResponse:
      type: object
      properties:
        token:
          type: string
          description: The generated token (only shown once)
        name:
          type: string
        type:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INVALID_ARGUMENT
              message: Invalid request parameters

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or missing token

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found